requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.5.2
bcrypt==4.3.0
cffi==1.17.1
click==8.1.8
cryptography==45.0.3
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.115.12
greenlet==3.1.1
h11==0.16.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
Mako==1.3.10
MarkupSafe==2.1.5
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.44.0
typing_extensions==4.13.2
uvicorn==0.33.0
zipp==3.20.2
